Refactored the dedicated nodes page structure to enhance maintainability. Relocated filters and associated functionality outside the nodes-table component. This separation aligns with best practices, ensuring that the nodes-table component solely focuses on rendering the table without any embedded logic. Work in progress.
